requirements.txt
asgiref>=3.6.0
Django>=4.0.8,<5.0
sqlparse>=0.4.3
psycopg2-binary>=2.9.3
netmiko>=4.1.2
django-bootstrap5>=22.2
pandas>=2.0.3
requests>=2.26.0
librouteros>=3.4.1
python-dotenv>=1.0.1
routeros-api
django-rq>=2.5.1
APScheduler>=3.10.0
redis>=4.5.4
rq>=1.15,<2.0
djangorestframework
djangorestframework-simplejwt
django-filter
whitenoise>=6.0.0
psutil>=5.9.0
prometheus-client>=0.16.0
